requirements.txt
numpy>=1.26
pandas>=2.2
pyyaml>=6.0
python-dotenv>=1.0
pyarrow>=15.0
scikit-learn>=1.4
torch>=2.2
tqdm>=4.66
tensorboard>=2.16
pytest>=8.0
